Rebase

+++ apps/pomodoro-backend/src/app/Modules/pomodoro/Services/Template/template.service.ts
@@ -1,17 +1,23 @@
 import { UpdateTemplateDTO } from './../../Dto/update-template-dto';
 import { SettingValueService } from './../SettingValue/settingvalue.service';
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { AddTemplateDTO } from '../../Dto/add-template-dto';
 import { PrismaService } from '../../../prisma/prisma.service';
 import { AddSettingDTO } from '../../Dto/add-setting-dto';
 import { TemplateFilter } from '../../Filters/TemplateFilter';
+import { TokenPayload } from '../../../auth/Services/authenticate.service';
+import { checkOwnerThrow } from '../../../auth/Guards/extract-payload.decorator';
 
 @Injectable()
 export class TemplateService {
-  DeleteTemplate(id: number) {
-     this.prisma.template.deleteMany({ where: { id: id } });
-  }
+  async DeleteTemplate(id: number) {
+    const template = await this.prisma.template.findFirst({
+      where: { id: id },
+    });
+    if (!template) return;
 
+    this.prisma.template.deleteMany({ where: { id: id } });
+  }
   async GetTemplate(id: number) {
     const template = await this.prisma.template.findFirst({
       where: { id: id },
@@ -33,9 +39,8 @@ export class TemplateService {
       },
     });
 
-    const defaultTemplate = (
-      await this.GetTemplateFilter({ userID: userID,default:true })
-    ).at(0);
+    const defaultTemplate = await this.GetUserDefault(userID);
+
     const copyValues: AddSettingDTO[] =
       defaultTemplate.template_SettingValue.map((settingValue) => {
         return <AddSettingDTO>{
@@ -58,12 +63,6 @@ export class TemplateService {
     });
   }
   async CreateUserDefault(userID: number) {
-
-    const defaultExists = (
-      await this.GetTemplateFilter({ userID: userID,default:true  })
-    ).at(0);
-    if(defaultExists) throw new BadRequestException("Default settings already exists for a user");
-
     const defaultTemplate = await this.prisma.template.create({
       data: {
         userOwner: { connect: { id: userID } },
@@ -86,6 +85,19 @@ export class TemplateService {
       },
     });
   }
+  async GetUserDefault(userID: number) {
+    const defaultExists = await this.prisma.template.findFirst({
+      where: {
+        userID: userID,
+        isDefault: true,
+      },
+      include: {
+        template_SettingValue: true,
+      },
+    });
+    if (defaultExists) return defaultExists;
+    return this.CreateUserDefault(userID);
+  }
   UpdateTemplate(id: number, dto: UpdateTemplateDTO) {
     return this.prisma.template.update({
       where: { id: id },
@@ -94,26 +106,40 @@ export class TemplateService {
       },
     });
   }
-  MapSettingTemplate(id:number,from:number,to:number)
-  {
+  async MapSettingTemplate(
+    id: number,
+    from: number,
+    to: number,
+    payload: TokenPayload
+  ) {
+    const fromSetting = await this.settingValueService.GetSetting(from);
+    const toSetting = await this.settingValueService.GetSetting(from);
+    const ownerFromTemplate = await this.GetTemplate(
+      fromSetting.ownerTemplateID
+    );
+    const ownerToTemplate = await this.GetTemplate(toSetting.ownerTemplateID);
+
+    checkOwnerThrow(ownerFromTemplate.userID, payload);
+    checkOwnerThrow(ownerToTemplate.userID, payload);
+
     return this.prisma.template.update({
       where: { id: id },
       data: {
-        template_SettingValue:{
-          disconnect:{id:from},
-          connect:{id:to}
-
-        }
+        template_SettingValue: {
+          disconnect: { id: from },
+          connect: { id: to },
+        },
+      },
+      include: {
+        template_SettingValue: true,
       },
     });
   }
 
-  GetTemplateFilter(filter: TemplateFilter) {
+  async GetTemplateFilter(filter: TemplateFilter) {
     return this.prisma.template.findMany({
       where: {
         userID: filter.userID,
-        id: filter.id,
-        isDefault: filter.default,
       },
       include: {
         template_SettingValue: true,
